refactor: export types from types.ts
